This gets rid of the one-thread-per-collection issue
It replaces it with a cached thread pool that makes sure tasks are all
executed in order.

I removed an executor.shutdown() call
at io.bluedb.disk.collection.BlueCollectionOnDisk.shutdown(BlueCollectionOnDisk.java:186).
Because cached thread pools clean themselves up this will not be an
issue unless you were relying on that shutdown to prevent future task
submissions, in which case you'll need to make CachedSingleThreadingPool
handle shutdowns.

The included test verifies that it is working as expected, but may
benefit from more scrutiny.