Index Condition Enhancements
- Collection Entity/Value Iterators now support index condition groups.
Group results are anded together, but conditions within each group can
be anded or orred.
- Added isInRange method to int and long index conditions
- Added a LongTimeKey class that can be indexed on when you want to
index a time in milliseconds. It needs to be handled differently since
it represents milliseconds instead of like an id. Mostly to avoid i-node
count getting out of hand and resource utilization going with it.