WhereKeyIsIn and Index Condition Optimization
- The query system could already support whitelisting what segments to
search if conditions can tell us that before executing the query. Now it
supports whitelisting what portion of each segment needs to be searched
since whereKeyIsIn and index conditions know exactly what keys it needs
to find. This helps to avoid having to search the whole segment when we
have the information to know that we can stop at a certain point.
- If we have the data to support it, the query system can now skip files
it knows it doesn't need to read or exit reading a file early if it
knows that it is done.