Added better support for active and completed time frame collections
- Added a new sub type of TimeKey called ActiveTimeKey. Active keys only
have a start time but version 2 collections will consider them active
and they will therefore be returned for any timeframe that includes time
at or after its start.
- The time query api now has methods for updating or replacing the key
and value of a time based record. The id and start time must remain the
same, but this allows you to set or change the end time of a record
without deleting and re-adding it to the collection.
- Timeframe keys no longer include the end time in the hash code and
equals. Time keys are considered the same if the time/start-time and id
is the same. The end time is important if we want to support the time
frame query feature for records that have a start and end, but the id
and start time are now considered the immutable, primary key. The end
time can change and it is still considered the same key.
- The Collection update and replace methods can now be used to update
the end time as long as you specify a new key that reflects the change
and which is equivalent to the original key.
- The time query api now has methods that allow you to include only
active or non-active records in a query.
- The IndividualChange class is now registered with FST, so that should
save a decent chunk of resources on pending change files.
- In conclusion, active and completed records can easily be stored in
the same collection now. Active records can be changed to completed
records and completed records can have their end time changed. The
existing query features will still function correctly and you can now
easily target active records in a query.
- Added an additional rollup size of 5 minutes to time segment size
settings. This will help to avoid batch upserts getting backed up by 7
seconds and creating an hour sized file chunk just because there isn't
another level between 6 seconds and one hour. Rolling up after changes
are done being added is nice for reads, but if we're still going to be
writing a lot to the current hour segment for awhile we don't want to
roll it up until later or else writes will suffer as we re-write the
entire hour segment data each time it gets updated.